Generate Rough Report of `EGRET` Outputs

Accepts the lists returned by `EGRET::modelEstimation` and `EGRET::runSeries` and generates a PDF report containing several coarse plots generated from `EGRET` functions for dealing with these lists. Useful for quick assessment of broad patterns in response but more finely-constructed plots should be built elsewhere

egret_report(
  eList_estim = NULL,
  eList_series = NULL,
  out_path = file.path(getwd(), "EGRET_report.pdf")
)

Arguments

eList_estim

(list) list returned by `EGRET::modelEstimation`

eList_series

(list) list returned by `EGRET::runSeries`

out_path

(character) name to save report as (must be a PDF) and (optionally) the path to export (defaults to `getwd()`)